@@ -24,7 +24,7 @@ repositories {
 // Apply a specific Java toolchain to ease working on different environments.
 java {
     toolchain {
-        languageVersion.set(JavaLanguageVersion.of(11))
+        languageVersion.set(JavaLanguageVersion.of(17))
     }
 }
 
@@ -39,15 +39,22 @@ idea {
 }
 
 val junitVersion: String by project
+val resilience4jVersion: String by project
 
 dependencies {
     api("org.slf4j:slf4j-api:2.0.13")
+    implementation("org.slf4j:slf4j-api:2.0.12")
+    implementation("com.fasterxml.jackson.core:jackson-databind:2.17.0")
+    implementation("com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.17.0")
+
+    implementation("io.github.resilience4j:resilience4j-all:$resilience4jVersion")
     // testing
     testRuntimeOnly("org.slf4j:slf4j-simple:2.0.13")
     testImplementation("org.junit.jupiter:junit-jupiter-api:$junitVersion")
     testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:$junitVersion")
     testRuntimeOnly("org.junit.platform:junit-platform-launcher:1.10.2")
     testImplementation("com.github.tomakehurst:wiremock-jre8:3.0.1")
+    testImplementation("org.hamcrest:hamcrest:2.2")
 }
 
 testing {
