@@ -2,8 +2,8 @@ package de.kleinkop.pushover4j;
 
 import java.time.LocalDateTime;
 
-public class ApplicationUsage {
-    private Integer limit;
-    private Integer remaining;
-    private LocalDateTime reset;
-}
+public record ApplicationUsage(
+    Integer limit,
+    Integer remaining,
+    LocalDateTime reset
+) {}
