@@ -3,17 +3,17 @@ package de.kleinkop.pushover4j;
 import java.io.File;
 import java.time.LocalDateTime;
 import java.time.OffsetDateTime;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.Set;
 
 public class Message {
-    private String message;
+    private final String message;
     private String title;
     private Priority priority = Priority.NORMAL;
     private String url;
     private String urlTitle;
-    private List<String> devices = new ArrayList<>();
+    private final Set<String> devices = new HashSet<>();
     private LocalDateTime timestamp;
     private boolean html = true;
     private String sound;
@@ -21,29 +21,20 @@ public class Message {
     private boolean monospace = false;
     private Integer retry;
     private Integer expire;
-    private List<String> tags = new ArrayList<>();
+    private final Set<String> tags = new HashSet<>();
 
     private Message(String msg) {
         this.message = msg;
     }
 
-    public Message(String message, String title, Priority priority, String url, String urlTitle, List<String> devices, LocalDateTime timestamp, boolean html, String sound, File image, boolean monospace, Integer retry, Integer expire) {
-        this.message = message;
-        this.title = title;
-        this.priority = priority;
-        this.url = url;
-        this.urlTitle = urlTitle;
-        this.devices = devices;
-        this.timestamp = timestamp;
-        this.html = html;
-        this.sound = sound;
-        this.image = image;
-        this.monospace = monospace;
-        this.retry = retry;
-        this.expire = expire;
+    private void validate() {
         if (priority == Priority.EMERGENCY) {
-            Objects.requireNonNull(retry, "Retry value required for emergency messages");
-            Objects.requireNonNull(expire, "Expiration value required for emergency messages");
+            if (retry == null || retry < 0) {
+                throw new IllegalArgumentException("Retry value required for emergency messages");
+            }
+            if (expire == null || expire < 0) {
+                throw new IllegalArgumentException("Expiration value required for emergency messages");
+            }
         }
 
         if (html && monospace) {
@@ -71,7 +62,7 @@ public class Message {
         return urlTitle;
     }
 
-    public List<String> getDevices() {
+    public Set<String> getDevices() {
         return devices;
     }
 
@@ -103,8 +94,8 @@ public class Message {
         return expire;
     }
 
-    public List<String> getTags() {
-        return tags;
+    public Set<String> getTags() {
+        return Collections.unmodifiableSet(tags);
     }
 
     public static Builder of(String msg) {
@@ -183,7 +174,9 @@ public class Message {
             return this;
         }
 
-        public Message build() {
+        public Message build()
+        {
+            template.validate();
             return template;
         }
     }
