@@ -7,7 +7,7 @@ public enum Priority {
     HIGH(1),
     EMERGENCY(2);
 
-    private int value;
+    private final int value;
 
     Priority(int value) {
         this.value = value;
